% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval_recursion}
\alias{eval_recursion}
\title{Evaluate an entire dynprog recursion.}
\usage{
eval_recursion(ranges, recursions)
}
\arguments{
\item{ranges}{The ranges specificatoin}

\item{recursions}{The recursions specification}
}
\value{
The filled out dynamic programming table
}
\description{
This function takes the \code{ranges} and \code{recursions} of a specification and
evaluate the dynprog expression, returning a filled out dynamic programming
table.
}
