% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_table_name}
\alias{get_table_name}
\title{Extract the table name from a pattern.}
\usage{
get_table_name(patterns)
}
\arguments{
\item{patterns}{The patterns used in the recursion.}
}
\value{
The table part of the pattern.
}
\description{
We generally assume that patterns are on the form \code{table[exprs]}
where \code{table} is the name of the dynamic programming table. This
function extract that name.
}
