% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{\%with\%}
\alias{\%with\%}
\title{Defines a table and connects it with a recursive formula.}
\usage{
tbl_name \%with\% recursion
}
\arguments{
\item{tbl_name}{A name we can use for a table. This name is used
in the \code{recursion} specification.}

\item{recursion}{Specification of a recursion that uses \code{tbl_name}.}
}
\value{
An object representing the recursion-expression.
This object is used to create the dynamic programming algorithm
when the \code{\%with\%} expression is combined with a \code{\%where\%}
expression that specifies the indices the algorithm should iterate over.
}
\description{
FIXME: more
}
\seealso{
\%where\%
}
