% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{make_pattern_match}
\alias{make_pattern_match}
\title{Translate a pattern into a predicate that checks the pattern.}
\usage{
make_pattern_match(pattern, range_vars)
}
\arguments{
\item{pattern}{An expression on the form \code{table[index-list]}}

\item{range_vars}{A list of the variables used in the ranges.}
}
\value{
An expression that tests \code{pattern} against \code{range_vars}.
}
\description{
Takes a pattern from the DSL and make a comparison of the
pattern specification against range variables.
}
