% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{make_recursion_case}
\alias{make_recursion_case}
\title{Construct a test for a case in the recursion}
\usage{
make_recursion_case(test_expr, value_expr, continue)
}
\arguments{
\item{test_expr}{The expression that must be true for the case to be applied}

\item{value_expr}{The value to compute if the test is true}

\item{continue}{The next case to check if this one isn't true}
}
\value{
An \code{if}-statement for checking and potentially evaluating one case.
}
\description{
This function creates an \code{if}-statement for testing if a case can be
applied.
}
