% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{make_update_expr}
\alias{make_update_expr}
\title{String together the case \code{if}-statements of a recursion.}
\usage{
make_update_expr(ranges, patterns, conditions, recursions)
}
\arguments{
\item{ranges}{The ranges specification}

\item{patterns}{The patterns specification}

\item{conditions}{The conditions specifications}

\item{recursions}{The recursions specification}
}
\value{
A series of \code{if}-\code{else}-statements for evaluating a recursion.
}
\description{
String together the case \code{if}-statements of a recursion.
}
