% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse}
\alias{parse}
\title{Function for parsing an entire dynprog expression.}
\usage{
parse(spec)
}
\arguments{
\item{spec}{The dynamic programming specification.}
}
\value{
\preformatted{An object containing the parsed specification.
}
}
\description{
This is the top-level parser. It will be invoked by \code{\link{\%where\%}}
once an entire expression has been parsed by the R parser. It is then
responsible for doing the DSL parsing before we evaluate the expression.
}
