% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_recursion}
\alias{parse_recursion}
\title{Parser for the recursion part of a specification.}
\usage{
parse_recursion(recursion)
}
\arguments{
\item{recursion}{The quosure wrapping the recursion of the specification.}
}
\value{
A parsed specification for recursions.
}
\description{
FIXME: more
}
