% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{\%where\%}
\alias{\%where\%}
\title{Connects a recursion with sequences it should recurse over.}
\usage{
recursion \%where\% ranges
}
\arguments{
\item{recursion}{Specification of the dynamic programming recursion.}

\item{ranges}{Specification of the index-ranges the recursion should
compute values over.}
}
\description{
FIXME: more
}
