% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_recursion}
\alias{parse_recursion}
\title{Parser for the recursion part of a specification.}
\usage{
parse_recursion(recursion)
}
\arguments{
\item{recursion}{The quosure wrapping the recursion of the specification.}
}
\value{
A parsed specification for recursions.
}
\description{
Parse the recursion part of an expressions.
}
\details{
The parser return a list with the following components:
\itemize{
\item \strong{recursion_env:}  The environment in which expressions should be
evaluated.
\item \strong{partterns:} A list of patterns, one per recursion case.
\item \strong{conditions:} A list of conditions, one per recursion case.
\item \strong{recursions:} A list of expressions, one per recursion case.
}
}
